Moved test out of src
