+++ test/sample-project/src/exports-sample.js
